update: update items pages OO
