+++ src/gui/task/task_create_page.py
@@ -2,7 +2,7 @@ import tkinter as tk
 from tkinter import ttk  
 from tkinter import messagebox
 from tkcalendar import DateEntry
-from src.gui.CRUD_base.base_create_page import BaseCreatePage
+from src.gui.base_CRUD.base_create_page import BaseCreatePage
 from src.gui.forms_base import EntryField, LabelCombobox, DescriptionText, DateField
 
 class TaskCreatePage(BaseCreatePage):
