+++ src/gui/task/updata_task_page.py
@@ -3,14 +3,12 @@ from tkinter import ttk
 from tkinter import messagebox
 from tkcalendar import DateEntry
 from src.gui.forms_base import EntryField, LabelCombobox, DescriptionText, DateField
+from src.gui.base_CRUD.base_updata_page import BaseUpdatePage
 
 
-class UpdateTaskPage(tk.Frame):
+class TaskUpdatePage(BaseUpdatePage):
     def __init__(self, master, mediator, controller, task):
-        super().__init__(master)
-        self.mediator = mediator
-        self.controler = controller
-        self.task = task
+        super().__init__(master, mediator, controller, task)
         self.create_widgets()
 
     def create_widgets(self):
@@ -22,43 +20,27 @@ class UpdateTaskPage(tk.Frame):
 
         priority = [ "Alta", "Media","Baixa"]
         
-        
         self.name_field = EntryField(self, "Nome:", entry_width, padding, self.mediator)
         self.priority_combobox = LabelCombobox(self, "Prioridade:", priority, entry_width, padding, self.mediator)
         self.end_date_field = DateField(self, "Data de Entrega:", entry_width, padding, self.mediator)
         self.notification_date_field = DateField(self, "Data da Notificacao:", entry_width, padding, self.mediator)
         self.description_text = DescriptionText(self, "Descrição:", 6, entry_width, padding, self.mediator)
         
-        self.name_field.set_value(self.task.name if self.task.name is not None else '')
-        self.priority_combobox.set_value(self.task.priority if self.task.priority is not None else '')
-        self.end_date_field.set_value(self.task.end_date)
-        self.notification_date_field.set_value(self.task.notification_date)
-        self.description_text.set_value(self.task.description if self.task.description is not None else '')
+        self.name_field.set_value(self.item.name if self.item.name is not None else '')
+        self.priority_combobox.set_value(self.item.priority if self.item.priority is not None else '')
+        self.end_date_field.set_value(self.item.end_date)
+        self.notification_date_field.set_value(self.item.notification_date)
+        self.description_text.set_value(self.item.description if self.item.description is not None else '')
        
-        # Criação de um frame adicional para os botões
-        button_frame = tk.Frame(self)
-        button_frame.pack(side="bottom", pady=10)
+        self.get_buttons()
 
-        # Botão de envio
-        submit_button = tk.Button(button_frame, text="Salvar", command=self.submit)
-        submit_button.grid(row=0, column=1, padx=5)
-
-        # Botão para fechar a janela
-        close_button = tk.Button(button_frame, text="Sair", command=lambda: self.controler.task_manager.open_task_page(self.task))
-        close_button.grid(row=0, column=0, padx=5)
-
-
-    def submit(self):
-        if not self.name_field.get_value():
-            messagebox.showerror("Erro", "O campo 'Nome' é obrigatório.")
-            return
-        task_data = {
+    def prepare_data(self):
+        data = {
             "name": self.name_field.get_value(),
             "priority": self.priority_combobox.get_value(),
             "end_date": self.end_date_field.get_value(),
             "notification_date": self.notification_date_field.get_value(),
             "description": self.description_text.get_value()
         }
-        self.mediator.notify(self, "submit", task_data)
-        self.master.destroy()
 
+        return data
